requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
firebase-admin==6.4.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
APScheduler==3.10.4
email-validator==2.1.0.post1
payos==1.1.0
python-zalo-bot==0.1.9
